# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/custom-error": {
      "get": {
        "operationId": "AppController_throwCustomError",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          }
        },
        "summary": "Cria um novo usuário",
        "tags": [
          "users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso"
          }
        },
        "summary": "Lista todos os usuários",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "summary": "Busca um usuário pelo ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário atualizado"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "summary": "Atualiza informações de um usuário",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário removido"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "summary": "Remove um usuário",
        "tags": [
          "users"
        ]
      }
    },
    "/errors/custom": {
      "get": {
        "operationId": "ErrorsController_getCustomError",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Usuário "
          }
        },
        "summary": "Dispara uma exceção personalizada",
        "tags": [
          "Errors"
        ]
      }
    },
    "/errors/bad-request": {
      "get": {
        "operationId": "ErrorsController_getBadRequest",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Parâmetros inválidos enviados"
          }
        },
        "summary": "Retorna erro 400 (Bad Request)",
        "tags": [
          "Errors"
        ]
      }
    },
    "/errors/not-found": {
      "get": {
        "operationId": "ErrorsController_getNotFound",
        "parameters": [],
        "responses": {
          "404": {
            "description": "O recurso solicitado não foi encontrado"
          }
        },
        "summary": "Retorna erro 404 (Not Found)",
        "tags": [
          "Errors"
        ]
      }
    },
    "/errors/unauthorized": {
      "get": {
        "operationId": "ErrorsController_getUnauthorized",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Usuário não autenticado"
          }
        },
        "summary": "Retorna erro 401 (Unauthorized)",
        "tags": [
          "Errors"
        ]
      }
    },
    "/errors/forbidden": {
      "get": {
        "operationId": "ErrorsController_getForbidden",
        "parameters": [],
        "responses": {
          "403": {
            "description": "Acesso negado para o usuário"
          }
        },
        "summary": "Retorna erro 403 (Forbidden)",
        "tags": [
          "Errors"
        ]
      }
    },
    "/errors/test/{id}": {
      "get": {
        "operationId": "ErrorsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recurso encontrado com sucesso"
          },
          "400": {
            "description": "O ID fornecido não é um número válido"
          },
          "401": {
            "description": "Usuário não autenticado"
          },
          "403": {
            "description": "Usuário autenticado, mas sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          }
        },
        "summary": "Dispara diferentes erros dependendo do parâmetro enviado",
        "tags": [
          "Errors"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Usuário registrado com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          }
        },
        "summary": "Registra um novo usuário",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        },
        "summary": "Realiza login com email e senha",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/perfil": {
      "get": {
        "operationId": "AuthController_getPerfil",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil retornado com sucesso"
          },
          "401": {
            "description": "Token ausente ou inválido"
          }
        },
        "summary": "Retorna o perfil do usuário autenticado (rota protegida)",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/me": {
      "get": {
        "operationId": "AdminController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin": {
      "get": {
        "operationId": "AdminController_getAdminData",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mensagem após logar como admin",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "API com Swagger",
    "description": "Documentação automática da API com Swagger",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "password": {
            "type": "string",
            "example": "12345678"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}